+++ src/commands/index.js
@@ -2,6 +2,7 @@
 
 module.exports = {
   createByTemplate: require('./createByTemplate'),
+  createFileByTemplate: require('./createFileByTemplate'),
   editTemplateGlobal: require('./editTemplateGlobal'),
   editTemplateWorkspace: require('./editTemplateWorkspace'),
 }